#13,#93[feat]홈 상단 뷰 페이저 개선
