+++ app/src/main/java/com/nbc/shownect/presentation/home/HorizontalMarginItemDecoration.kt
@@ -1,0 +1,21 @@
+package com.nbc.shownect.presentation.home
+
+import android.content.Context
+import android.graphics.Rect
+import android.view.View
+import androidx.annotation.DimenRes
+import androidx.recyclerview.widget.RecyclerView
+
+//리사이클러뷰 항목에 수평 마진을 추가
+class HorizontalMarginItemDecoration(context: Context, @DimenRes horizontalMarginInDp: Int) :
+    RecyclerView.ItemDecoration() {
+    private val horizontalMarginInPx: Int =
+        context.resources.getDimension(horizontalMarginInDp).toInt()
+
+    override fun getItemOffsets(
+        outRect: Rect, view: View, parent: RecyclerView, state: RecyclerView.State
+    ) {
+        outRect.right = horizontalMarginInPx
+        outRect.left = horizontalMarginInPx
+    }
+}
